Split out collections module.
