+++ ethereum/core/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
   implementation project(':bls')
   implementation project(':data')
+  implementation project(':infrastructure:collections')
   implementation project(':infrastructure:io')
   implementation project(':infrastructure:metrics')
   implementation project(':data:signingrecord')
