+++ ethereum/core/src/main/java/tech/pegasys/teku/core/stategenerator/CachingTaskQueue.java
@@ -28,8 +28,8 @@ import org.hyperledger.besu.plugin.services.metrics.Counter;
 import org.hyperledger.besu.plugin.services.metrics.LabelledMetric;
 import tech.pegasys.teku.infrastructure.async.AsyncRunner;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.collections.LimitedMap;
 import tech.pegasys.teku.infrastructure.metrics.TekuMetricCategory;
-import tech.pegasys.teku.util.collections.LimitedMap;
 
 public class CachingTaskQueue<K, V> {
