+++ ethereum/core/src/main/java/tech/pegasys/teku/core/stategenerator/StateCache.java
@@ -18,7 +18,7 @@ import java.util.Map;
 import java.util.Optional;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.datastructures.state.BeaconState;
-import tech.pegasys.teku.util.collections.LimitedMap;
+import tech.pegasys.teku.infrastructure.collections.LimitedMap;
 
 class StateCache {
   private final Map<Bytes32, BeaconState> cache;
