+++ ethereum/datastructures/build.gradle
@@ -5,6 +5,7 @@ dependencies {
 
   implementation project(':bls')
   implementation project(':infrastructure:async')
+  implementation project(':infrastructure:collections')
   implementation project(':infrastructure:logging')
   implementation project(':pow')
   implementation project(':util')
