+++ ethereum/datastructures/src/main/java/tech/pegasys/teku/datastructures/state/TransitionCaches.java
@@ -17,10 +17,10 @@ import java.util.List;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.bls.BLSPublicKey;
+import tech.pegasys.teku.infrastructure.collections.cache.Cache;
+import tech.pegasys.teku.infrastructure.collections.cache.LRUCache;
+import tech.pegasys.teku.infrastructure.collections.cache.NoOpCache;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.util.cache.Cache;
-import tech.pegasys.teku.util.cache.LRUCache;
-import tech.pegasys.teku.util.cache.NoOpCache;
 
 /** The container class for all transition caches. */
 public class TransitionCaches {
