+++ ethereum/statetransition/build.gradle
@@ -1,11 +1,12 @@
 dependencies {
   implementation project(':bls')
   implementation project(':data')
-  implementation project(':infrastructure:metrics')
   implementation project(':ethereum:core')
   implementation project(':ethereum:datastructures')
   implementation project(':infrastructure:async')
+  implementation project(':infrastructure:collections')
   implementation project(':infrastructure:logging')
+  implementation project(':infrastructure:metrics')
   implementation project(':infrastructure:subscribers')
   implementation project(':pow')
   implementation project(':protoarray')
