+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/util/FutureItems.java
@@ -22,8 +22,8 @@ import java.util.concurrent.ConcurrentSkipListMap;
 import java.util.function.Function;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import tech.pegasys.teku.infrastructure.collections.LimitedSet;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.util.collections.LimitedSet;
 
 /** Holds items with slots that are in the future relative to our node's current slot */
 public class FutureItems<T> {
