+++ infrastructure/collections/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+  implementation 'com.google.guava:guava'
+}
