+++ infrastructure/collections/src/main/java/tech/pegasys/teku/infrastructure/collections/LimitedMap.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.collections;
+package tech.pegasys.teku.infrastructure.collections;
 
 import com.google.common.cache.CacheBuilder;
 import java.util.Map;
