+++ infrastructure/collections/src/main/java/tech/pegasys/teku/infrastructure/collections/cache/NoOpCache.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.cache;
+package tech.pegasys.teku.infrastructure.collections.cache;
 
 import java.util.Optional;
 import java.util.function.Function;
