+++ infrastructure/collections/src/test/java/tech/pegasys/teku/infrastructure/collections/LimitedMapTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.collections;
+package tech.pegasys.teku.infrastructure.collections;
 
 import static org.assertj.core.api.Assertions.assertThat;
