+++ networking/eth2/build.gradle
@@ -2,12 +2,13 @@ dependencies {
   api project(':networking:p2p')
   implementation project(':bls')
   implementation project(':data')
-  implementation project(':infrastructure:metrics')
   implementation project(':ethereum:core')
   implementation project(':ethereum:datastructures')
   implementation project(':ethereum:statetransition')
   implementation project(':infrastructure:async')
+  implementation project(':infrastructure:collections')
   implementation project(':infrastructure:logging')
+  implementation project(':infrastructure:metrics')
   implementation project(':infrastructure:subscribers')
   implementation project(':ssz')
   implementation project(':storage')
