+++ settings.gradle
@@ -15,6 +15,7 @@ include 'data:yaml'
 include 'fuzz'
 include 'infrastructure'
 include 'infrastructure:async'
+include 'infrastructure:collections'
 include 'infrastructure:events'
 include 'infrastructure:io'
 include 'infrastructure:metrics'
