+++ storage/build.gradle
@@ -1,11 +1,12 @@
 dependencies {
   implementation project(':bls')
-  implementation project(':infrastructure:metrics')
   implementation project(':ethereum:core')
   implementation project(':ethereum:datastructures')
   implementation project(':infrastructure:async')
+  implementation project(':infrastructure:collections')
   implementation project(':infrastructure:io')
   implementation project(':infrastructure:logging')
+  implementation project(':infrastructure:metrics')
   implementation project(':pow')
   implementation project(':protoarray')
   implementation project(':storage:api')
