+++ sync/build.gradle
@@ -1,13 +1,14 @@
 dependencies {
     implementation project(':data')
-    implementation project(':infrastructure:metrics')
     implementation project(':ethereum:core')
     implementation project(':ethereum:datastructures')
     implementation project(':ethereum:statetransition')
     implementation project(':infrastructure:async')
     implementation project(':networking:p2p')
     implementation project(':networking:eth2')
+    implementation project(':infrastructure:collections')
     implementation project(':infrastructure:logging')
+    implementation project(':infrastructure:metrics')
     implementation project(':infrastructure:subscribers')
     implementation project(':services:serviceutils')
     implementation project(':ssz')
